Pulled apart the github service and commented out some testing temporarily
